[FIXED] API reservas

+++ digital-booking-backend/src/main/java/com/grupo9/digitalBooking/music/model/service/BookingService.java
@@ -2,6 +2,8 @@ package com.grupo9.digitalBooking.music.model.service;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.grupo9.digitalBooking.music.model.DTO.BookingDTO;
+import com.grupo9.digitalBooking.music.model.DTO.BookingResponseDTO;
+import com.grupo9.digitalBooking.music.model.DTO.UserResponseDTO;
 import com.grupo9.digitalBooking.music.model.entities.Instrument;
 import com.grupo9.digitalBooking.music.model.entities.Status;
 import com.grupo9.digitalBooking.music.model.entities.UserApp;
@@ -22,6 +24,9 @@ import java.util.Set;
 @Service
 public class BookingService implements IBookingService {
 
+    private final String NOT_AVAILABLE = "no disponible";
+    private final String AVAILABLE = "no disponible";
+
     @Autowired
     private IBooking bookingRepository;
     
@@ -41,13 +46,19 @@ public class BookingService implements IBookingService {
         return bookingRepository.save(booking);
     }
 
+    private Boolean existById(Long id) {
+        return bookingRepository.findById(id).isPresent();
+    }
     @Override
-    public Booking createBooking(BookingDTO bookingDTO) {
-        UserApp user = userRepository.findUserById(bookingDTO.getUserApp().getId());
+    public BookingResponseDTO createBooking(BookingDTO bookingDTO) {
+        UserApp user = userRepository.findUserById(bookingDTO.getUser().getId());
         Booking newBooking = new Booking();
+        BookingResponseDTO response = new BookingResponseDTO();
 
         if (user != null) {
             Instrument instrument = instrumentRepository.findInstrumentById(bookingDTO.getInstrument().getId());
+            Status notAvailable = mapper.convertValue(statusRepository.findByName(NOT_AVAILABLE), Status.class);
+
             Status status = statusRepository.findStatusById(bookingDTO.getStatus().getId());
 
             newBooking.setStartDate(bookingDTO.getStartDate());
@@ -55,42 +66,67 @@ public class BookingService implements IBookingService {
             newBooking.setInstrument(instrument);
             newBooking.setStatus(status);
             newBooking.setUserApp(user);
+
+            instrument.setStatus(notAvailable);
+            instrumentRepository.save(instrument);
             newBooking = saveBooking(newBooking);
 
-            instrument.setStatus(status);
-            instrumentRepository.save(instrument);
+            response = mapper.convertValue(newBooking, BookingResponseDTO.class);
+            response.setUser(mapper.convertValue(newBooking.getUserApp(), UserResponseDTO.class));
         }
 
-        return newBooking;
+
+        return response;
+
+
+
     }
 
     @Override
-    public BookingDTO readBooking(Long id) {
+    public BookingResponseDTO readBooking(Long id) {
         Optional<Booking> bookings = bookingRepository.findById(id);
         BookingDTO bookingDTO = null;
-        if(bookings.isPresent())
+        BookingResponseDTO response = new BookingResponseDTO();
+        if(bookings.isPresent()) {
             bookingDTO = mapper.convertValue(bookings, BookingDTO.class);
+            response = mapper.convertValue(bookingDTO, BookingResponseDTO.class);
+        }
 
-        return bookingDTO;
+        return response;
     }
 
     @Override
-    public void modifyBooking(BookingDTO bookingDTO) {
-//        saveBooking(bookingDTO);
+    public BookingResponseDTO modifyBooking(BookingDTO bookingDTO) {
+        BookingResponseDTO bookingResponseDTO = null;
+        Boolean valditateBooking = existById(bookingDTO.getId());
+
+        if(valditateBooking) {
+            Booking booking =  saveBooking(mapper.convertValue(bookingDTO, Booking.class));
+            bookingResponseDTO = mapper.convertValue(booking, BookingResponseDTO.class);
+        }
+
+        return bookingResponseDTO;
+
     }
 
     @Override
-    public void removeBooking(Long id) {
-        bookingRepository.deleteById(id);
+    public Boolean removeBooking(Long id) {
+        Boolean response = false;
+        Boolean exist = existById(id);
+        if(exist) {
+            bookingRepository.deleteById(id);
+            response = true;
+        }
+        return response;
     }
 
     @Override
-    public Set<BookingDTO> getAll() {
+    public Set<BookingResponseDTO> getAll() {
         List<Booking> bookings = bookingRepository.findAll();
-        Set<BookingDTO> bookingDTOS = new HashSet<>();
+        Set<BookingResponseDTO> bookingDTOS = new HashSet<>();
 
         for (Booking booking1 : bookings) {
-            bookingDTOS.add(mapper.convertValue(booking1, BookingDTO.class));
+            bookingDTOS.add(mapper.convertValue(booking1, BookingResponseDTO.class));
         }
 
         return bookingDTOS;
